Put dismissible message at start of message
This is to switch around the self::$dismissButtonFormat to the front of the alert box as is shown on getbootstrap.com. In the current order, the X will be in the wrong place if the alert contains a display:block element such as the render output of the FlashMessenger helper.