Fix "formElement" container config.
From what I understand, aliases are processed before factories, so when
merging the configuration of all ZF modules, the 'aliases/formElement'
entry in zend-form takes precedence and the factory entry
'factories/formElement' is ignored.

In order to override the zend-form's element, we need to create an alias
with that name.